+++ Kernel/Arch/x86/Time/HPET.h
