+++ Kernel/Arch/x86/common/PCSpeaker.cpp
@@ -5,8 +5,8 @@
  */
 
 #include <Kernel/Arch/x86/IO.h>
+#include <Kernel/Arch/x86/Time/PIT.h>
 #include <Kernel/Arch/x86/common/PCSpeaker.h>
-#include <Kernel/Time/PIT.h>
 
 void PCSpeaker::tone_on(int frequency)
 {
