+++ Kernel/CMakeLists.txt
@@ -298,11 +298,6 @@ set(KERNEL_SOURCES
     Thread.cpp
     ThreadBlockers.cpp
     ThreadTracer.cpp
-    Time/APICTimer.cpp
-    Time/HPET.cpp
-    Time/HPETComparator.cpp
-    Time/PIT.cpp
-    Time/RTC.cpp
     Time/TimeManagement.cpp
     TimerQueue.cpp
     UBSanitizer.cpp
@@ -344,6 +339,12 @@ if ("${SERENITY_ARCH}" STREQUAL "i686" OR "${SERENITY_ARCH}" STREQUAL "x86_64")
         Arch/x86/PCI/Controller/HostBridge.cpp
         Arch/x86/PCI/IDELegacyModeController.cpp
         Arch/x86/PCI/Initializer.cpp
+
+        Arch/x86/Time/APICTimer.cpp
+        Arch/x86/Time/HPET.cpp
+        Arch/x86/Time/HPETComparator.cpp
+        Arch/x86/Time/PIT.cpp
+        Arch/x86/Time/RTC.cpp
 
         Arch/x86/VGA/IOArbiter.cpp
     )
